权限

+++ app/system/biz/infra/do/role.go
@@ -10,6 +10,6 @@ type Role interface {
 	Update(req *auth.RoleInfo) error
 	Delete(id int64) error
 	RoleInfoByID(ID int64) (roleInfo *auth.RoleInfo, err error)
-	List(req *base.PageInfoReq) (roleInfoList []*auth.RoleInfo, total int, err error)
+	List(req *base.PageInfoReq) (roleInfoList []*auth.RoleInfo, total int64, err error)
 	UpdateStatus(ID int64, status int64) error
 }
